# services/auth-service/docs/swagger.yaml
basePath: /
definitions:
  handler.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  handler.LoginResponse:
    properties:
      access_token:
        type: string
      access_token_expires_at:
        type: string
      id:
        type: integer
      refresh_token:
        type: string
      refresh_token_expires_at:
        type: string
      user:
        $ref: '#/definitions/handler.UserResponse'
    type: object
  handler.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    type: object
  handler.RefreshTokenResponse:
    properties:
      access_token:
        type: string
      access_token_expires_at:
        type: string
    type: object
  handler.RevokeTokenRequest:
    properties:
      refresh_token:
        type: string
    type: object
  handler.UserResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      role:
        type: string
      username:
        type: string
    type: object
host: localhost:3001
info:
  contact: {}
  description: API documentation for the Chat Room Authentication Service
  title: Chat Room Auth API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate and return tokens
      parameters:
      - description: Login request body
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Login user
      tags:
      - auth
  /logout:
    post:
      description: Logout by deleting the refresh token cookie and entry in database
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - auth
  /refresh-token:
    post:
      consumes:
      - application/json
      description: Refresh the access token using a valid refresh token
      parameters:
      - description: Refresh token request body
        in: body
        name: refreshTokenRequest
        required: true
        schema:
          $ref: '#/definitions/handler.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RefreshTokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Refresh access token
      tags:
      - auth
  /revoke-token:
    post:
      consumes:
      - application/json
      description: Revoke a refresh token to invalidate future access tokens
      parameters:
      - description: Revoke token request body
        in: body
        name: revokeTokenRequest
        required: true
        schema:
          $ref: '#/definitions/handler.RevokeTokenRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Revoke a refresh token
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    description: '"JWT Authorization header using the Bearer scheme. Example: \"Bearer
      {token}\""'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"